% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{acc_location}
\alias{acc_location}
\title{Location}
\usage{
acc_location(lat, lon, radius = 50, pages = 1)
}
\arguments{
\item{lat, lon}{Coordinates to retrieve local news articles from.}

\item{radius}{Radius in miles. from coordinates, maximum 300 miles.}

\item{pages}{Number of pages of results to collect (100 per page).}
}
\description{
This endpoint retrieves latest local news articles around a specific city/location.
}
\examples{
\dontrun{
 local <- acc_location(27.964157, -82.452606, pages = 2)
}

}
