% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{articles}
\alias{articles}
\alias{acc_local}
\alias{acc_world}
\title{Articles}
\usage{
acc_local(lat, lon, popular = FALSE, radius = 50, pages = 1)

acc_world(q = NULL, popular = FALSE, pages = 1)
}
\arguments{
\item{lat, lon}{Coordinates to retrieve local news articles from.}

\item{popular}{Set to \code{TRUE} to retrieve popular local or world news.}

\item{radius}{Radius in miles. from coordinates, maximum 300 miles.}

\item{pages}{Number of pages of results to collect (100 per page).}

\item{q}{Query to search for world latest articles.}
}
\description{
This endpoint retrieves latest or most popular local or world news articles around a specific city/location.
}
\examples{
\dontrun{
 world <- acc_world(popular = TRUE)
 local <- acc_local(27.964157, -82.452606, pages = 2)
}

}
