% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticate}
\alias{authenticate}
\alias{accunews_key}
\title{Setup}
\usage{
accunews_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://www.accunewsapp.com/api/v1/docs/#authentication}.}
}
\description{
Setup your session, all subsequent calls will be done using the API key.
}
\note{
You can specify \code{ACCUNEWS_API_KEY} as environment variable, likely in your \code{.Renviron} file.
}
\examples{
\dontrun{
accunews_key("xXXxxXxXxXXx")  
}

}
