% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{source}
\alias{source}
\alias{acc_local_sources}
\alias{acc_world_sources}
\title{Sources}
\usage{
acc_local_sources(lat, lon, radius = 50, pages = 1)

acc_world_sources(popular = FALSE, type = NULL, pages = 1)
}
\arguments{
\item{lat}{Coordinates to retrieve local news articles from.}

\item{lon}{Coordinates to retrieve local news articles from.}

\item{radius}{Radius in miles. from coordinates, maximum 300 miles.}

\item{pages}{Number of pages of results to collect (100 per page).}

\item{popular}{Set to \code{TRUE} to retrieve popular local or world news.}

\item{type}{Type of source/publisher to query for e.g. \code{politics} or \code{sports}.}
}
\description{
Get local and world news sources.
}
\examples{
\dontrun{
 src <- acc_world_sources(type = "business")
}

}
